Back/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
deprecation==2.1.0
flask==3.1.0
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.1.0
openai==1.64.0
packaging==24.2
postgrest==0.19.3
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.0
redis==5.2.1
returns==0.24.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.38
storage3==0.11.3
strenum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typing-extensions==4.12.2
websockets==14.2
yarl==1.18.3
